+++ Sources/InstanaAgent/Configuration/InstanaConfiguraton.swift
@@ -17,8 +17,6 @@ import Foundation
 
 // Use a reference type to avoid a copy when having concurrency
 class InstanaConfiguration {
-    static let maxSlowSendInterval = 3600.0
-
     enum SuspendReporting {
         /// Reporting is suspended while the device battery is low.
         case lowBattery
@@ -48,8 +46,6 @@ class InstanaConfiguration {
     }
 
     struct Defaults {
-        // 0 seconds means disable network probing
-        static let slowSendInterval: Instana.Types.Seconds = 0.0
         static let reporterSendDebounce: Instana.Types.Seconds = 2.0
         static let reporterSendLowBatteryDebounce: Instana.Types.Seconds = 10.0
         static let gzipReport = ProcessInfo.ignoreZIPReporting ? false : true
@@ -67,6 +63,7 @@ class InstanaConfiguration {
     var suspendReporting: Set<SuspendReporting>
     var monitorTypes: Set<MonitorTypes>
     var slowSendInterval: Instana.Types.Seconds
+    var usiRefreshTimeIntervalInHrs: Double
     var reporterSendDebounce: Instana.Types.Seconds
     var reporterSendLowBatteryDebounce: Instana.Types.Seconds
     var maxRetries: Int
@@ -78,7 +75,8 @@ class InstanaConfiguration {
     var isValid: Bool { !key.isEmpty && !reportingURL.absoluteString.isEmpty }
 
     required init(reportingURL: URL, key: String, httpCaptureConfig: HTTPCaptureConfig,
-                  enableCrashReporting: Bool, slowSendInterval: Double) {
+                  enableCrashReporting: Bool, slowSendInterval: Instana.Types.Seconds,
+                  usiRefreshTimeIntervalInHrs: Double) {
         self.reportingURL = reportingURL
         self.key = key
         self.httpCaptureConfig = httpCaptureConfig
@@ -88,6 +86,7 @@ class InstanaConfiguration {
             monitorTypes.insert(.crash)
         }
         self.slowSendInterval = slowSendInterval
+        self.usiRefreshTimeIntervalInHrs = usiRefreshTimeIntervalInHrs
         reporterSendDebounce = Defaults.reporterSendDebounce
         reporterSendLowBatteryDebounce = Defaults.reporterSendLowBatteryDebounce
         maxRetries = Defaults.maxRetries
@@ -99,8 +98,11 @@ class InstanaConfiguration {
     }
 
     static func `default`(key: String, reportingURL: URL, httpCaptureConfig: HTTPCaptureConfig = .automatic,
-                          enableCrashReporting: Bool, slowSendInterval: Double = 0.0) -> InstanaConfiguration {
+                          enableCrashReporting: Bool, slowSendInterval: Instana.Types.Seconds = 0.0,
+                          usiRefreshTimeIntervalInHrs: Double = defaultUsiRefreshTimeIntervalInHrs)
+        -> InstanaConfiguration {
         self.init(reportingURL: reportingURL, key: key, httpCaptureConfig: httpCaptureConfig,
-                  enableCrashReporting: enableCrashReporting, slowSendInterval: slowSendInterval)
+                  enableCrashReporting: enableCrashReporting, slowSendInterval: slowSendInterval,
+                  usiRefreshTimeIntervalInHrs: usiRefreshTimeIntervalInHrs)
     }
 }