+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.6.5
 - Add crash to mobile feature list and send to Instana backend
+- Add user_session_id (usi) to beacons
 
 ## 1.6.4
 - Fix issue for setCaptureHeaders method not capture http response headers